requirements.txt
# core
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
tqdm>=4.0

# autogen + openai
pyautogen[openai]>=0.2
openai>=1.0

# api
flask>=3.0
flask-cors>=4.0

# dev
ruff>=0.4
pytest>=8.0
pytest-asyncio>=0.23
autogen